Cleanup imports and distinguish contexts
