+++ frontend/src/main.jsx
@@ -1,17 +1,19 @@
+import "bootstrap/dist/css/bootstrap.min.css";
+import "./styles/main.css";
+
 import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
 import { BrowserRouter, Route, Routes } from "react-router";
-import { data } from "./features/data.js";
+
 import ModeWrap from "./components/modewrap.jsx";
-import "bootstrap/dist/css/bootstrap.min.css";
-import "./styles/main.css";
-import MainArea from "./components/mainarea.jsx";
-import Accolade from "./components/accolade.jsx";
-import Identity from "./components/identity.jsx";
-import AccoList from "./components/accolist.jsx";
-import Mistaken from "./components/mistaken.jsx";
-import Recently from "./components/recently.jsx";
-import UserPast from "./components/userpast.jsx";
+import Accolade from "./routes/accolade.jsx";
+import AccoList from "./routes/accolist.jsx";
+import Identity from "./routes/identity.jsx";
+import MainArea from "./routes/mainarea.jsx";
+import Mistaken from "./routes/mistaken.jsx";
+import Recently from "./routes/recently.jsx";
+import UserPast from "./routes/userpast.jsx";
+import { data } from "./features/data.js";
 
 createRoot(document.getElementById("root")).render(
   <Provider store={data}>
