+++ frontend/src/routes/mainarea.jsx
@@ -1,7 +1,8 @@
-import Navigate from "./navigate.jsx";
+import { Container } from "react-bootstrap";
 import { Outlet } from "react-router";
-import Container from "react-bootstrap/Container";
-import LoadNote from "./loadnote.jsx";
+
+import LoadNote from "../components/loadnote.jsx";
+import Navigate from "../components/navigate.jsx";
 
 export default function MainArea() {
   return (
